+++ flv_sbe/sbe_messages/src/messages/client_messages/client_login/sbe_decode.rs
@@ -11,7 +11,7 @@ pub fn decode_client_login_message(buffer: &[u8]) -> SbeResult<ClientLoginMessag
     csg = csg.header(header);
 
     let sbe_message_type = csg.message_type();
-    let message_type = MessageType::from(sbe_message_type as u8);
+    let message_type = MessageType::from(sbe_message_type as u16);
     assert_eq!(message_type, MessageType::ClientLogin);
 
     let client_id = csg.client_id();
