+++ flv_sbe/sbe_messages/src/messages/data_messages/start_data/sbe_decode.rs
@@ -13,7 +13,7 @@ pub fn decode_start_data_message(buffer: &[u8]) -> SbeResult<StartDataMessage> {
     csg = csg.header(header);
 
     let sbe_message_type = csg.message_type();
-    let message_type = MessageType::from(sbe_message_type as u8);
+    let message_type = MessageType::from(sbe_message_type as u16);
     assert_eq!(message_type, MessageType::StartData);
 
     let client_id = csg.client_id();
