+++ flv_sbe/sbe_messages/src/messages/data_messages/stop_all_data/sbe_decode.rs
@@ -14,7 +14,7 @@ pub fn decode_stop_all_data_message(buffer: &[u8]) -> SbeResult<StopAllDataMessa
     csg = csg.header(header);
 
     let sbe_message_type = csg.message_type();
-    let message_type = MessageType::from(sbe_message_type as u8);
+    let message_type = MessageType::from(sbe_message_type as u16);
     assert_eq!(message_type, MessageType::StopAllData);
 
     let client_id = csg.client_id();
